backend/blockchain_service/certificate_verification_system/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
web3==6.3.0
qrcode==7.4.2
pillow==10.0.0
opencv-python-headless==4.8.0.76
pyzbar==0.1.8
pytesseract==0.3.10
python-multipart==0.0.6
pytest==7.4.2
python-dotenv==1.0.0
